python-rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LangChain & AI
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.11
langchain-openai==0.0.5

# Vector Database
pinecone-client==3.0.2

# Embeddings
sentence-transformers==2.2.2
huggingface-hub==0.19.4

# LLM
openai==1.3.5

# Data Processing
numpy==1.26.2
pandas==2.1.3

# Text Processing
korean-text-processors==0.4.4
konlpy==0.6.0

# HTTP Client
httpx==0.25.1
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Redis
redis==5.0.1

# Development
python-multipart==0.0.6
